refactor: 대시보드의 불필요 컨테이너 제거
대시보드의 탭을 통합하면서 불필요해진 기존 컨테이너들을 삭제함.
