+++ client/src/views/DashBoardPage/index.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from 'react';
 import TransactionModal from '@container/TransactionModal';
 import ModalToggleButton from '@components/common/buttons/ModalToggleButton';
 import MainLayout from '@/components/common/layouts/MainLayout';
-import DashboardContainer from '@/container/DashboardMain';
+import DashboardContainer from '@/container/DashboardContainer';
 
 const DashBoardPage = (): JSX.Element => {
   const [showModal, setShowModal] = useState(false);
